requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.10.0
pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.28.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
xmltodict>=0.13.0
opencv-python>=4.8.0
wandb>=0.15.0
xformers>=0.0.20
